+++ armada_backend/api_ship.py
@@ -12,6 +12,7 @@ from armada_command import armada_api
 from armada_command.consul.consul import consul_query, consul_put
 from runtime_settings import override_runtime_settings
 from utils import get_ship_name, get_other_ship_ips, get_current_datacenter
+from armada_command.consul import kv
 
 
 def _get_current_consul_mode():
@@ -72,7 +73,9 @@ class Name(api_base.ApiCommand):
         ship_name, error = self.get_post_parameter('name')
         if error:
             return self.status_error(error)
-        if not ship_name or ship_name == 'None':
+        other_ship_names = [get_ship_name(ip) for ip in get_other_ship_ips()]
+        name_taken = ship_name in other_ship_names
+        if not ship_name or ship_name == 'None' or name_taken:
             return self.status_error('Incorrect ship name: {}'.format(ship_name))
         set_ship_name(ship_name)
         return self.status_ok()
@@ -83,6 +86,10 @@ class Join(api_base.ApiCommand):
         consul_host, error = self.get_post_parameter('host')
         if error:
             return self.status_error(error)
+
+        ship = get_ship_name()
+        local_services = kv.kv_list('ships/{}/service/'.format(ship)) or []
+        local_services_data = {key: kv.kv_get(key) for key in local_services}
 
         armada_size = _get_armada_size()
         if armada_size > 1:
@@ -108,6 +115,9 @@ class Join(api_base.ApiCommand):
             supervisor_server = xmlrpclib.Server('http://localhost:9001/RPC2')
             hermes_init_output = supervisor_server.supervisor.startProcessGroup('hermes_init')
             get_logger().info('hermes_init start: {}'.format(hermes_init_output))
+            set_ship_name(ship)
+            for key, data in local_services_data.items():
+                kv.kv_set(key, data)
             return self.status_ok()
         return self.status_error('Waiting for armada restart timed out.')