+++ armada_command/command_diagnose.py
@@ -39,8 +39,8 @@ def command_diagnose(args):
             if status == 'recovering':
                 params = instance['params']
                 print('RESTART_CONTAINER_PARAMETERS:')
-                print(params)
-            elif status == 'crashed':
+                print(json.dumps(params, indent=4, sort_keys=True))
+            elif status in ['crashed', 'not-recovered']:
                 params = instance['params']
                 print('RESTART_CONTAINER_PARAMETERS:')
                 print(json.dumps(params, indent=4, sort_keys=True))